requirements.txt
streamlit~=1.51.0
crewai~=1.6.1
python-dotenv~=1.2.1
requests~=2.32.5
pydantic~=2.12.5
langchain_openai~=1.1.0